feat(extension): refactor Networks component as a container
