+++ projects/extension/src/components/MenuContent.tsx
@@ -1,19 +1,14 @@
 import React from "react"
-import { MenuItemContent } from "./MenuItemContent"
 
-interface MenuContentProps {
-  activeMenu: number
+type MenuContentProps = {
   children: React.ReactNode
 }
 
-export const MenuContent = ({ activeMenu = 0, children }: MenuContentProps) => (
+export const MenuContent = ({ children }: MenuContentProps) => (
   <div className="relative flex flex-col w-full min-w-0 mb-6 break-words rounded">
     <div className="flex-auto px-4 py-5">
       <div className="tab-content tab-space">
-        {React.Children.map(
-          children,
-          (c, i) => activeMenu === i && <MenuItemContent child={c} index={i} />,
-        )}
+        <div className="block">{children}</div>
       </div>
     </div>
   </div>
