--- projects/extension/src/components/MenuItemContent.tsx
@@ -1,12 +1,0 @@
-import { ReactNode } from "react"
-
-interface MenuItemContentProps {
-  index: number
-  child: ReactNode
-}
-
-export const MenuItemContent = ({ child, index }: MenuItemContentProps) => (
-  <div className={"block"} id={`"${"link".concat(index.toString())}"`}>
-    {child}
-  </div>
-)
