+++ projects/extension/src/components/Networks.tsx
@@ -1,16 +1,35 @@
+import { useMemo } from "react"
+import { useActiveChains } from "../hooks/useActiveChains"
 import { NetworkTabProps } from "../types"
 import NetworkTab from "./NetworkTab"
 
-interface NetworksProps {
-  networks: NetworkTabProps[]
-}
-
-export const Networks = ({ networks }: NetworksProps) => {
+export const Networks = () => {
+  const chains = useActiveChains()
+  const networks: NetworkTabProps[] = useMemo(
+    () =>
+      chains.map(({ chainName, isWellKnown, details }) => {
+        return {
+          isWellKnown,
+          name: chainName,
+          health: {
+            isSyncing: details[0].isSyncing,
+            peers: details[0].peers,
+            status: "connected",
+            bestBlockHeight: details[0].bestBlockHeight,
+          },
+          apps: details.map(({ url }) => ({
+            name: url ?? "",
+            url: url,
+          })),
+        }
+      }),
+    [chains],
+  )
   return (
     <section className="mx-0 md:mx-12 xl:mx-36 2xl:mx-64">
       <div className="font-inter font-bold text-3xl pb-4">Networks</div>
       {networks.length ? (
-        networks.map((network: NetworkTabProps, i: number) => {
+        networks.map((network, i) => {
           const { name, health, apps, isWellKnown } = network
           return (
             <NetworkTab
